Update the proxy HTML rewrite logic to avoid building a complete parse tree, instead using the lower level token stream API. Doing so removes the need to create "missing" HTML elements, reducing the collateral changes to the rewritten HTML.

Resolves: #9766

### Benchmarks
The new approach is slightly faster, and scales better with large documents. The documents tested were the kubernetes dashboard index page ("dash", 2.3K), a snapshot of a gmail tab ("mail", 1.7M), and a snapshot of the one-page HTML spec ("html", 9.4M).

**This change**
```
BenchmarkDash	   10000	    123018 ns/op	   26512 B/op	     408 allocs/op
BenchmarkMail	      50	  24182171 ns/op	 5638125 B/op	   64658 allocs/op
BenchmarkHtml	       3	 356755878 ns/op	67479413 B/op	  901099 allocs/op
```

**Previous method**
```
BenchmarkDash	   10000	    138123 ns/op	   24320 B/op	     367 allocs/op
BenchmarkMail	      50	  25806537 ns/op	 5770287 B/op	   35871 allocs/op
BenchmarkHtml	       2	 576100073 ns/op	76078952 B/op	  779853 allocs/op
```