@timstclair 2015-06-16
FYI, merged with master so this PR no longer includes the commit from #9756