@lavalamp 2015-06-16
This is great! Performance improvements are in scope for 1.0. Just two tiny nits.