@timstclair 2015-06-16
Thanks Dan. Should I squash the fixes?